+++ src/index.html
@@ -6,13 +6,13 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <title>Mesto</title>
-  <link rel="stylesheet" href="./pages/index.css">
+  <link rel="stylesheet" href="<%=htmlWebpackPlugin.files.chunks.main.css %>">
 </head>
 
 <body>
   <div class="root">
     <header class="header root__section">
-      <img src="./images/logo.svg" alt="mesto logo" class="logo">
+      <img src="<%=require('./images/logo.svg')%>" alt="mesto logo" class="logo">
     </header>
     <div class="profile root__section">
       <div class="user-info">
@@ -27,24 +27,10 @@
     </div>
     <div class="places-list root__section">
 
-      <!-- Пример вёрстки карточки: -->
-      <!-- 
-      <div class="place-card">
-        <div class="place-card__image" style="background-image: url(https://pictures.s3.yandex.net/frontend-developer/cards-compressed/kamchatka.jpg)">
-          <button class="place-card__delete-icon"></button>
-        </div>
-        <div class="place-card__description">
-          <h3 class="place-card__name">Камчатка</h3>
-          <button class="place-card__like-icon"></button>
-        </div>
-      </div>-->
-
-      <!-- конец примера. -->
-
     </div>
     <div class="popup">
       <div class="popup__content">
-        <img src="./images/close.svg" alt="" class="popup__close">
+        <img src="<%=require('./images/close.svg')%>" alt="" class="popup__close">
         <h3 class="popup__title">Новое место</h3>
         <form class="popup__form" name="new">
           <input type="text" name="title" class="popup__input popup__input_type_name" placeholder="Название"
@@ -62,7 +48,7 @@
 
     <div class="popup profile-popup">
       <div class="popup__content">
-        <img src="./images/close.svg" alt="" class="popup__close popup__close_type_edit">
+        <img src="<%=require('./images/close.svg')%>" alt="" class="popup__close popup__close_type_edit">
         <h3 class="popup__title">Редактировать профиль</h3>
         <form class="popup__form" name="edit">
           <input type="text" name="name" class="popup__input popup__input_type_name" placeholder="Имя" minlength="2"
@@ -80,22 +66,12 @@
 
     <div class="popup image-popup">
       <div class="image-popup__content">
-        <img src="./images/close.svg" alt="" class="image-popup__close">
+        <img src="<%=require('./images/close.svg')%>" alt="" class="image-popup__close">
         <img src="https://pictures.s3.yandex.net/frontend-developer/cards-compressed/arkhyz.jpg" alt=""
           class="image-popup__image">
       </div>
     </div>
-    <script src="js/api.js"></script>
-    <script src="js/Owner.js"></script>
-    <script src="js/initialCards.js"></script>
-    <script src="js/UserInfo.js"></script>
-    <script src="js/Card.js"></script>
-    <script src="js/CardList.js"></script>
-    <script src="js/Popup.js"></script>
-    <script src="js/PopupPlace.js"></script>
-    <script src="js/PopupEdit.js"></script>
-    <script src="js/PopupImage.js"></script>
-    <script src="./index.js"></script>
+    <script src="<%= htmlWebpackPlugin.files.chunks.main.entry %>"></script>
 
 </body>
