--- src/js/Owner.js
@@ -1,5 +1,0 @@
-class Owner {
-    constructor(id) {
-      this.id = id;
-    }
-  }
