+++ src/modules/Owner.js
@@ -1,0 +1,5 @@
+export default class Owner {
+  constructor(id) {
+    this.id = id;
+  }
+}
