+++ src/modules/UserInfo.js
@@ -1,5 +1,5 @@
-class UserInfo {
-    constructor( api, owner) {
+export default class UserInfo {
+    constructor(api, owner) {
         this.name = document.querySelector(".user-info__name");
         this.about = document.querySelector(".user-info__job");
         this.avatar = document.querySelector(".user-info__photo");
@@ -8,7 +8,7 @@ class UserInfo {
     }
 
     getUserInfo(path) {
-        this.api.get(path).then(user =>{
+        this.api.get(path).then(user => {
             this.owner.id = user._id;
 
             this.name.textContent = user.name;
